// src/app/models/parametersCreate.ts
export class ParametersCreate{
        age: number|string = '';
        birthday: string = '';
        growth: number|string = '';
        clothsize: number|string = '';
        chestsize: number|string = '';
        waistsize: number|string = '';
        thighsize: number|string = '';
        footsize: number|string = '';
        necksize: number|string = '';
        headsize: number|string = '';
        tatu: boolean|string = '';
        piercing: boolean|string = '';
        appearancetype: number|string = '';
        haircolor: number|string = '';
        city: number|string = '';
        eyecolor: number|string = '';
        gender: number|string = '';
        employmenttype: any = '';
}